+++ htdocs/emails.php
@@ -39,6 +39,12 @@ if (!Access::canAccessAssociateEmails(Auth::getUserID())) {
     exit;
 }
 
+// accept prj_id from GET to ease administration (link bookmarking)
+$prj_id = isset($_GET['prj_id']) ? (int) $_GET['prj_id'] : null;
+if ($prj_id && $prj_id != Auth::getCurrentProject()) {
+    AuthCookie::setProjectCookie($prj_id);
+}
+
 $pagerRow = Support::getParam('pagerRow');
 if (empty($pagerRow)) {
     $pagerRow = 0;
@@ -56,7 +62,7 @@ $list = Support::getEmailListing($options, $pagerRow, $rows);
 $tpl->assign('list', $list['list']);
 $tpl->assign('list_info', $list['info']);
 $tpl->assign('issues', Issue::getColList());
-$tpl->assign('accounts', Email_Account::getAssocList(Auth::getCurrentProject()));
+$tpl->assign('accounts', Email_Account::getAssocList($prj_id));
 
 $prefs = Prefs::get(Auth::getUserID());
 $tpl->assign('refresh_rate', $prefs['email_refresh_rate'] * 60);